# project/openapi.yaml
openapi: 3.0.3
info:
  title: 我的数字老师 - 后端接口
  version: 0.1.0
  description: |
    由当前 uni-app 前端项目整理出的"需要/预留"的后端接口规格。
    可直接导入 Swagger / Apifox / Postman。
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        ok:
          type: boolean
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "ok"
        data:
          nullable: true
    RegisterRequest:
      type: object
      required: [name, account, password]
      properties:
        name:
          type: string
          example: 小明
        account:
          type: string
          example: test@example.com
        phone:
          type: string
          nullable: true
          example: "13800000000"
        password:
          type: string
          format: password
          example: "123456"
    RegisterResponseData:
      type: object
      properties:
        userId:
          type: string
          example: u_123
    LoginRequest:
      type: object
      required: [account, password]
      properties:
        account:
          type: string
          example: test@example.com
        password:
          type: string
          format: password
          example: "123456"
    LoginResponseData:
      type: object
      properties:
        accessToken:
          type: string
          example: <jwt_or_any_token>
        expiresIn:
          type: integer
          example: 7200
        user:
          type: object
          properties:
            id:
              type: string
              example: u_123
            name:
              type: string
              example: 小明
            account:
              type: string
              example: test@example.com
    MeResponseData:
      type: object
      properties:
        id:
          type: string
          example: u_123
        name:
          type: string
          example: 张三
        account:
          type: string
          example: 1008611
    AccountItem:
      type: object
      properties:
        id:
          type: string
          example: a1
        name:
          type: string
          example: 小朋友
        account:
          type: string
          example: 1008611
    AccountsListResponseData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountItem'
    AccountsCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "新账号名"
    AccountsSwitchRequest:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: string
          example: a1
    AccountsSwitchResponseData:
      type: object
      properties:
        activeAccountId:
          type: string
          example: a1
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: 注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RegisterResponseData'
  /auth/login:
    post:
      tags: [Auth]
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponseData'
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MeResponseData'
  /accounts:
    get:
      tags: [Accounts]
      summary: 获取账号列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountsListResponseData'
    post:
      tags: [Accounts]
      summary: 创建账号
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountsCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountItem'
  /accounts/{accountId}/delete:
    post:
      tags: [Accounts]
      summary: 删除账号
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
  /accounts/switch:
    post:
      tags: [Accounts]
      summary: 切换当前账号
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountsSwitchRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountsSwitchResponseData'
...